# firewatch-frontend/openapi.json
{
  "components": {
    "schemas": {
      "ActionQueueItem": {
        "properties": {
          "days_overdue": {
            "title": "Days Overdue",
            "type": "integer"
          },
          "due_date": {
            "title": "Due Date",
            "type": "string"
          },
          "kind": {
            "enum": [
              "review",
              "response"
            ],
            "title": "Kind",
            "type": "string"
          },
          "risk_id": {
            "title": "Risk Id",
            "type": "string"
          },
          "risk_title": {
            "title": "Risk Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "risk_id",
          "risk_title",
          "due_date",
          "days_overdue"
        ],
        "title": "ActionQueueItem",
        "type": "object"
      },
      "ActionQueueResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionQueueItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "ActionQueueResponse",
        "type": "object"
      },
...
      "DashboardSummaryResponse": {
        "properties": {
          "by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Severity",
            "type": "object"
          },
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Status",
            "type": "object"
          },
          "overdue_responses": {
            "title": "Overdue Responses",
            "type": "integer"
          },
          "overdue_reviews": {
            "title": "Overdue Reviews",
            "type": "integer"
          },
          "risk_matrix": {
            "items": {
              "items": {
                "type": "integer"
              },
              "type": "array"
            },
            "title": "Risk Matrix",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "by_status",
          "by_severity",
          "overdue_responses",
          "overdue_reviews",
          "risk_matrix"
        ],
        "title": "DashboardSummaryResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarkAllReadResponse": {
        "properties": {
          "marked": {
            "title": "Marked",
            "type": "integer"
          }
        },
        "required": [
          "marked"
        ],
        "title": "MarkAllReadResponse",
        "type": "object"
      },
      "NotificationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "unread_total": {
            "title": "Unread Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "unread_total"
        ],
        "title": "NotificationListResponse",
        "type": "object"
      },
      "NotificationResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "link": {
            "title": "Link",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "risk_human_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Human Id"
          },
          "risk_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/NotificationType"
          }
        },
        "required": [
          "id",
          "type",
          "risk_id",
          "risk_human_id",
          "title",
          "message",
          "link",
          "created_at",
          "read_at"
        ],
        "title": "NotificationResponse",
        "type": "object"
      },
      "NotificationType": {
        "enum": [
          "risk_assigned",
          "review_overdue",
          "response_overdue",
          "risk_changed"
        ],
        "title": "NotificationType",
        "type": "string"
      },
...
      "ScoreHistoryPoint": {
        "properties": {
          "avg_score": {
            "title": "Avg Score",
            "type": "number"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "date": {
            "title": "Date",
            "type": "string"
          }
        },
        "required": [
          "date",
          "avg_score",
          "count"
        ],
        "title": "ScoreHistoryPoint",
        "type": "object"
      },
      "ScoreHistoryResponse": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/ScoreHistoryPoint"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "points"
        ],
        "title": "ScoreHistoryResponse",
        "type": "object"
      },
      "ScoreTotalsBySeverityPoint": {
        "properties": {
          "critical": {
            "title": "Critical",
            "type": "integer"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "high": {
            "title": "High",
            "type": "integer"
          },
          "low": {
            "title": "Low",
            "type": "integer"
          },
          "medium": {
            "title": "Medium",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "ScoreTotalsBySeverityPoint",
        "type": "object"
      },
      "ScoreTotalsBySeverityResponse": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/ScoreTotalsBySeverityPoint"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "points"
        ],
        "title": "ScoreTotalsBySeverityResponse",
        "type": "object"
      },
      "UnreadCountResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count"
        ],
        "title": "UnreadCountResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Cybersecurity Risk Register API -- NIST 800-30 aligned",
    "title": "Firewatch",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/dashboard/action-queue": {
      "get": {
        "operationId": "get_action_queue_api_dashboard_action_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionQueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Action Queue",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/dashboard/score-history": {
      "get": {
        "operationId": "get_score_history_api_dashboard_score_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "format": "date",
              "title": "End",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Score History",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/dashboard/score-totals-by-severity": {
      "get": {
        "operationId": "get_score_totals_by_severity_api_dashboard_score_totals_by_severity_get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "format": "date",
              "title": "End",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreTotalsBySeverityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Score Totals By Severity",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/dashboard/summary": {
      "get": {
        "operationId": "get_dashboard_summary_api_dashboard_summary_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dashboard Summary",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/api/notifications": {
      "get": {
        "operationId": "list_notifications_api_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/mark-all-read": {
      "post": {
        "operationId": "mark_all_read_api_notifications_mark_all_read_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkAllReadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark All Read",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/unread-count": {
      "get": {
        "operationId": "get_unread_count_api_notifications_unread_count_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadCountResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Unread Count",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/{notification_id}/read": {
      "post": {
        "operationId": "mark_one_read_api_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "title": "Notification Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark One Read",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  }
}